move constants to the root folder
