+++ www/src/mocks/hackernews_handlers.ts
@@ -3,7 +3,7 @@ import {
     MOCK_COMMENT_ID,
     MOCK_SERVER_URL,
     MOCK_STORY_ID,
-} from "@/lib/constants";
+} from "@/constants";
 import { range } from "@/lib/range";
 import { rest } from "msw";
